+++ src/plugin/Ls_Plugin.c
@@ -7,12 +7,55 @@
 #include "Ls_Plugin.h"
 
 
+static char** create_exec_argv(
+        int argc,
+        char* argv[],
+        List options,
+        Error* error) {
+
+    List exec_argv_list = List_literal(Array_List, error, "ls", NULL);
+    Error discard;
+
+    if (*error) {
+        return NULL;
+    }
+
+    if (options != NULL) {
+        List_extend(exec_argv_list, options, error);
+
+        if (*error) {
+            List_delete(exec_argv_list, &discard);
+            return NULL;
+        }
+    }
+
+    for (int i = 0; i <= argc; ++i) {
+        List_add(exec_argv_list, (intptr_t) argv[i], error);
+
+        if (*error) {
+            List_delete(exec_argv_list, &discard);
+            return NULL;
+        }
+    }
+
+    char** exec_argv = (char**) List_to_array(exec_argv_list, sizeof(char*), error);
+    List_delete(exec_argv_list, &discard);
+
+    if (*error) {
+        return NULL;
+    }
+
+    *error = NULL;
+    return exec_argv;
+}
+
+
 static int exec_forkpty(char* file, char* argv[], Error* error) {
     int saved_stderr = dup(STDERR_FILENO);
 
     if (saved_stderr == -1) {
         *error = strerror(errno);
-        return PLUGIN_INVALID_FD_OUT;
+        return -1;
     }
 
     int child_fd_out;
@@ -20,7 +63,7 @@ static int exec_forkpty(char* file, char* argv[], Error* error) {
 
     if (child_pid == -1) {
         *error = strerror(errno);
-        return PLUGIN_INVALID_FD_OUT;
+        return -1;
     }
     else if (child_pid != 0) {
         close(saved_stderr);
@@ -30,12 +73,12 @@ static int exec_forkpty(char* file, char* argv[], Error* error) {
 
     if (dup2(saved_stderr, STDERR_FILENO) == -1) {
         *error = strerror(errno);
-        return PLUGIN_INVALID_FD_OUT;
+        return -1;
     }
 
     if (execvp(file, argv) == -1) {
         *error = strerror(errno);
-        return PLUGIN_INVALID_FD_OUT;
+        return -1;
     }
     else {
         exit(EXIT_SUCCESS);
@@ -53,47 +96,38 @@ static const char* get_name() {
 }
 
 
-static int run(int fd_in, int argc, char** argv, List options, Error* error) {
-    if (fd_in != PLUGIN_INVALID_FD_OUT) {
+static List run(
+        List fds_in,
+        int argc,
+        char* argv[],
+        List options,
+        Error* error) {
+
+    if ((List_length(fds_in) > 0) && (argc == 0)) {
         *error = NULL;
-        return fd_in;
+        return fds_in;
     }
 
-    List ls_argv = List_literal(Array_List, error, "ls", NULL);
-    Error discard;
+    char** exec_ls_argv = create_exec_argv(argc, argv, options, error);
 
     if (*error) {
-        return PLUGIN_INVALID_FD_OUT;
-    }
-
-    if (options != NULL) {
-        List_extend(ls_argv, options, error);
-
-        if (*error) {
-            List_delete(ls_argv, &discard);
-            return PLUGIN_INVALID_FD_OUT;
-        }
-    }
-
-    for (int i = 0; i <= argc; ++i) {
-        List_add(ls_argv, (intptr_t) argv[i], error);
-
-        if (*error) {
-            List_delete(ls_argv, &discard);
-            return PLUGIN_INVALID_FD_OUT;
-        }
-    }
-
-    char** exec_ls_argv = (char**) List_to_array(ls_argv, sizeof(char*), error);
-    List_delete(ls_argv, &discard);
-
-    if (*error) {
-        return PLUGIN_INVALID_FD_OUT;
+        return NULL;
     }
 
     int fd_out = exec_forkpty(exec_ls_argv[0], exec_ls_argv, error);
     free(exec_ls_argv);
-    return fd_out;
+
+    if (*error) {
+        return NULL;
+    }
+
+    List_add(fds_in, (intptr_t) fd_out, error);
+
+    if (*error) {
+        return NULL;
+    }
+
+    return fds_in;
 }